final changes commited

+++ src/main/java/parser/insertParser.java
@@ -22,7 +22,6 @@ public class insertParser {
 
 	Lock lock = new ReentrantLock();
 	public void insertQuery(String query) throws IOException {
-		// String[] queryValues = query.split(" ");
 		String colNames = "";
 		String values = "";
 		Matcher m = Pattern.compile(regex.BETWEENBRACKETS).matcher(query);
@@ -45,6 +44,8 @@ public class insertParser {
 		String[] columns = colNames.trim().split(",");
 		int fileExists = 1;
 		int size = 0;
+		int fkcheck= 1;
+		int pkcheck = 1;
 		long startTime = System.nanoTime();
 		if (entries.length != columns.length) {
 			System.err.println("Column names don't match the column values");
@@ -53,22 +54,26 @@ public class insertParser {
 			FailInsertionRecord failInsertionRecord = new FailInsertionRecord();
 			failInsertionRecord.event("",timeElapsed);
 		} else {
+			
 			for (int i = 0; i < columns.length; i++) {
-				lock.lock();
 				BufferedReader in = new BufferedReader(new FileReader(metaTablePath));
 				while ((line = in.readLine()) != null) {
-					if (line.contains(columns[i].trim()) && line.toLowerCase().contains("primary--key")) {
+					if (!output.contains(columns[i])&&line.contains(columns[i].trim()) && line.toLowerCase().contains("primary--key")) {
 						if (checkPrimaryConstraint(tablePath, entries[i].trim()) == true) {
 							output += "#" + columns[i] + "=" + entries[i] + ", ";
+		
 							size++;
 						} else {
-							System.err.println("Primary key constraint fails");
+							if(pkcheck==1) {
+								System.err.println("Primary key constraint fails");
+								
+							}
+							pkcheck=0;
 						}
 
 					} else if (line.contains(columns[i].trim()) && line.toLowerCase().contains("foreign--key")) {
 						String[] references = line.toLowerCase().split("--references--");
 						String[] refTableName = references[1].split("\\(");
-						// System.out.println(refTableName[0]);
 						String refPath = Database.getDatabase() + "/" + refTableName[0] + ".txt";
 						File file = new File(refPath);
 						if (!file.exists()) {
@@ -77,11 +82,16 @@ public class insertParser {
 							}
 							
 							fileExists=0;
-						} else if (checkPrimaryConstraint(refPath, columns[i].trim()) == false && file.exists()) {
+						} else if (!output.contains(columns[i])&& checkPrimaryConstraint(refPath, columns[i].trim()) == false && file.exists()) {
 							output += columns[i] + "=" + entries[i] + ", ";
 							size++;
 						} else {
-							System.err.println("Foreign key constraint fails");
+							if(fkcheck==1) {
+								System.err.println("Foreign key constraint fails");
+								
+							}
+							fkcheck=0;
+							
 						}
 
 					}
@@ -90,20 +100,23 @@ public class insertParser {
 							&& !line.toLowerCase().contains("primary--key")
 							&& !line.toLowerCase().contains("foreign--key")) {
 						output += columns[i] + "=" + entries[i] + ", ";
+						
 						size++;
 					}
-
 				}
-				FileWriter fileWriter = null;
-				if (size == columns.length && fileExists==1) {
+				
+
+			}
+			FileWriter fileWriter = null;
+			if (size == columns.length && fileExists==1 && fkcheck==1 && pkcheck==1) {
 					fileWriter = new FileWriter(tablePath, true);
 					fileWriter.write(System.lineSeparator());
 					fileWriter.write(output);
-					lock.unlock();
+					
 					System.out.println("Row successfully inserted.");
+					
 					fileWriter.flush();
-				}
-
+				
 			}
 			long endTime = System.nanoTime();
 			long timeElapsed = endTime - startTime;
